Add validation to category
