+++ app/assets/stylesheets/nav.css.scss
@@ -1,0 +1,13 @@
+@import "bootstrap/dist/css/bootstrap";
+
+.navbar {
+  padding-left: 40px;
+  padding-right: 40px;
+  border-bottom: 3px solid #ed5e36;
+}
+
+.navbar-brand {
+  font-weight: bold;
+  letter-spacing: -1px;
+  padding-right: 100px;
+}
