+++ app/helpers/category_helper.rb
@@ -1,0 +1,3 @@
+module CategoryHelper
+  
+end
